requirements.txt
# Django для веб-CRM
Django==4.2.7
django-environ==0.11.2
psycopg2-binary==2.9.9

# Telegram Bot
aiogram==3.3.0
aiohttp==3.9.1

# Утилиты
python-dotenv==1.0.0

# REST API
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Production server
gunicorn==21.2.0

